Even Liber Abbaci

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	unsigned c, i, j, k, s;
+	unsigned long c, i, j, k, s;
 
 	i = s = 0;
 	j = i;
